backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
websockets==12.0
langchain>=0.2.0
langgraph>=0.2.0
langchain-xai>=0.1.0
httpx>=0.27.0
python-multipart>=0.0.9
pydantic>=2.7.0
python-dotenv>=1.0.0
